Implement gameplay balance tuning and workflow updates
