--- CON
@@ -1,5 +1,0 @@
-import { Game } from "./src/Game.js";
-
-const canvas = document.getElementById("game");
-const game = new Game(canvas);
-game.start();
