--- TASK_BOARD.md
@@ -1,11 +1,0 @@
-# Task Board: Floor Exit Rework
-
-The floor-exit rework is implemented and documented in [README.md](README.md).
-
-## Open Follow-Ups
-- None currently tracked for this initiative.
-
-## Validation Commands
-- `npm run check`
-- `npm run validate:boss`
-- `npm run perf:test`
