+++ docs/TASK_BOARD.md
@@ -1,0 +1,18 @@
+# Task Board
+
+Branch: `feature/<branch-name-here>`
+
+Use this file as the working board for the current branch. Keep tasks finite, testable, and tied to concrete validation steps. When a feature is complete, roll the lasting summary into the relevant project docs and reset this file again.
+
+## Active Tasks
+- [ ] Add the next scoped task here.
+  Scope:
+  Validation:
+
+## Follow-Ups
+- None recorded.
+
+## Validation Commands
+- `npm run check`
+- `npm run validate:boss`
+- `npm run perf:test`
