+++ src/game/enemySpawnFactories.js
@@ -1,0 +1,188 @@
+export function spawnGhost(game, x, y) {
+  const hp = game.rollScaledEnemyHealth(game.config.enemy.ghostHpMin, game.config.enemy.ghostHpMax);
+  const speed = 85 + Math.random() * 35;
+  return {
+    type: "ghost",
+    x,
+    y,
+    size: 20,
+    speed,
+    hp,
+    maxHp: hp,
+    baseMaxHp: hp,
+    baseSpeed: speed,
+    hpBarTimer: 0,
+    damageMin: game.config.enemy.ghostDamageMin,
+    damageMax: game.config.enemy.ghostDamageMax,
+    baseDamageMin: game.config.enemy.ghostDamageMin,
+    baseDamageMax: game.config.enemy.ghostDamageMax,
+    contactAttackCooldown: 0
+  };
+}
+
+export function spawnTreasureGoblin(game, x, y) {
+  const hp = game.rollScaledEnemyHealth(game.config.enemy.goblinHpMin, game.config.enemy.goblinHpMax);
+  return {
+    type: "goblin",
+    x,
+    y,
+    size: 16,
+    speed: 95,
+    hp,
+    maxHp: hp,
+    hpBarTimer: 0,
+    damageMin: game.config.enemy.goblinDamageMin,
+    damageMax: game.config.enemy.goblinDamageMax,
+    goldEaten: 0,
+    aggression: 0.12,
+    wanderAngle: Math.random() * Math.PI * 2,
+    wanderTimer: 0.5 + Math.random() * 0.8
+  };
+}
+
+export function spawnAnimatedArmor(game, x, y) {
+  const hp = game.rollScaledEnemyHealth(game.config.enemy.armorHpMin, game.config.enemy.armorHpMax);
+  return {
+    type: "armor",
+    x,
+    y,
+    size: 24,
+    speed: game.config.enemy.armorSpeed,
+    hp,
+    maxHp: hp,
+    hpBarTimer: 0,
+    damageMin: game.config.enemy.armorDamageMin,
+    damageMax: game.config.enemy.armorDamageMax
+  };
+}
+
+export function spawnMimic(game, x, y) {
+  const hp = game.rollScaledEnemyHealth(game.config.enemy.mimicHpMin, game.config.enemy.mimicHpMax);
+  return {
+    type: "mimic",
+    x,
+    y,
+    homeX: x,
+    homeY: y,
+    size: 20,
+    speed: game.config.enemy.mimicSpeed,
+    hp,
+    maxHp: hp,
+    hpBarTimer: 0,
+    damageMin: game.config.enemy.mimicDamageMin,
+    damageMax: game.config.enemy.mimicDamageMax,
+    dormant: true,
+    revealed: false,
+    tongueCooldown: 0,
+    tongueTimer: 0,
+    tongueDirX: 1,
+    tongueDirY: 0,
+    tongueLength: 0
+  };
+}
+
+export function spawnRatArcher(game, x, y) {
+  const hp = game.rollScaledEnemyHealth(game.config.enemy.ratArcherHpMin, game.config.enemy.ratArcherHpMax);
+  return {
+    type: "rat_archer",
+    x,
+    y,
+    size: 20,
+    speed: game.config.enemy.ratArcherSpeed,
+    hp,
+    maxHp: hp,
+    hpBarTimer: 0,
+    damageMin: game.config.enemy.ratArcherContactDamageMin,
+    damageMax: game.config.enemy.ratArcherContactDamageMax,
+    rangedDamageMin: game.config.enemy.ratArcherDamageMin,
+    rangedDamageMax: game.config.enemy.ratArcherDamageMax,
+    dirX: 1,
+    dirY: 0,
+    shotWindupTimer: 0,
+    shotIntervalTimer: 0,
+    burstCooldownTimer: 0,
+    dodgeCooldownTimer: 0,
+    dodgeTimer: 0,
+    dodgeVx: 0,
+    dodgeVy: 0,
+    shotsRemaining: game.config.enemy.ratArcherBurstShots,
+    coverTargetX: x,
+    coverTargetY: y,
+    repositionTimer: 0
+  };
+}
+
+export function spawnSkeletonWarrior(game, x, y) {
+  const hp = game.rollScaledEnemyHealth(game.config.enemy.skeletonWarriorHpMin, game.config.enemy.skeletonWarriorHpMax);
+  return {
+    type: "skeleton_warrior",
+    x,
+    y,
+    size: 20,
+    speed: game.config.enemy.skeletonWarriorSpeed,
+    hp,
+    maxHp: hp,
+    baseMaxHp: hp,
+    baseSpeed: game.config.enemy.skeletonWarriorSpeed,
+    hpBarTimer: 0,
+    damageMin: game.config.enemy.skeletonWarriorDamageMin,
+    damageMax: game.config.enemy.skeletonWarriorDamageMax,
+    baseDamageMin: game.config.enemy.skeletonWarriorDamageMin,
+    baseDamageMax: game.config.enemy.skeletonWarriorDamageMax,
+    attackCooldown: 0,
+    contactAttackCooldown: 0,
+    collapsed: false,
+    collapseTimer: 0,
+    reviveAtEnd: false
+  };
+}
+
+export function spawnNecromancer(game, x, y) {
+  const hpBase = game.rollScaledEnemyHealth(game.config.enemy.necromancerHpMin, game.config.enemy.necromancerHpMax);
+  const hpMultiplier = Number.isFinite(game.config.enemy.necromancerBossHpMultiplier)
+    ? Math.max(1, game.config.enemy.necromancerBossHpMultiplier)
+    : 1;
+  const speedMultiplier = Number.isFinite(game.config.enemy.necromancerBossSpeedMultiplier)
+    ? Math.max(1, game.config.enemy.necromancerBossSpeedMultiplier)
+    : 1;
+  const hp = Math.max(1, Math.round(hpBase * hpMultiplier));
+  const speed = game.config.enemy.necromancerSpeed * speedMultiplier;
+  return {
+    type: "necromancer",
+    x,
+    y,
+    size: 28,
+    speed,
+    hp,
+    maxHp: hp,
+    baseMaxHp: hp,
+    baseSpeed: speed,
+    hpBarTimer: 9999,
+    damageMin: game.config.enemy.necromancerDamageMin,
+    damageMax: game.config.enemy.necromancerDamageMax,
+    isFloorBoss: true,
+    strafeDir: Math.random() < 0.5 ? -1 : 1,
+    strafeTimer: 1.1 + Math.random() * 1.2,
+    summonCooldown: 0,
+    castCooldown: 0
+  };
+}
+
+export function spawnSkeleton(game, x, y, options = {}) {
+  const hp = game.rollScaledEnemyHealth(game.config.enemy.skeletonHpMin, game.config.enemy.skeletonHpMax);
+  return {
+    type: "skeleton",
+    x,
+    y,
+    size: 18,
+    speed: game.config.enemy.skeletonSpeed,
+    hp,
+    maxHp: hp,
+    hpBarTimer: 0,
+    damageMin: game.config.enemy.skeletonDamageMin,
+    damageMax: game.config.enemy.skeletonDamageMax,
+    summonedByNecromancer: !!options.summonedByNecromancer,
+    summonerBoss: !!options.summonerBoss,
+    summonLife: Number.isFinite(options.summonLife) ? options.summonLife : null
+  };
+}
